HACK: Wait briefly after clear_costmap_recovery
There appears to be a race condition wherein move_base will still fail right
after clear_cost_map_recovery finishes due to what must be a race condition.
This delay hacks around it.

However, it also has a bona fide purpose: It gives the sensors a bit of time
to repopulate the cost map with obstacles that are currently visible, which is
safer than beginning to move right away.