move_base: Adds plan persistence to wait for transient obstacles to clear
Solves this problem: If something blocks the path to the current goal, the
robot would abandon it immediately, switching to a new global plan, even
though that may route us all the way around a huge gondola. Then, we could get
80% through that long path when we get blocked again, so it would turn back
the old way.

Sometimes this would go back and forth for a very long time.

Now, whenever the local planner needs a new global plan, we will accept it
right away if it is not much longer than the previous plan.  But if it is a
lot longer, we will stop and replan over and over until we get a plan similar
in length to the previous one or until a timeout has expires.

This prevents transient obstacles from always sending us around the long way.

TODO: Add dynamic reconfigure params; Make the delay scale in proportion to
the extra length of the proposed new plan.