amcl: add gaussian motion model
This model uses integrated odometry data at the odometry rate to add
noise to the particle filter. It is more accurate than either of the
other models already in AMCL in the case when laser scans update
infrequently. When sensor data updates at a much lower rate than
odometry, the bearing calculation in the diff and omni models is very
incorrect in understanding how the robot actually moved.