Add support for unknown space tracking
NO_INFORMATION was being tracked as LETHAL in several places, as if
we never track unknown.  Now we determine if the costmap is setup
for unknown cell tracking, and then determine if we should behave in
as if unknown is lethal.

There are a few new lines like this in the code:
allow_unknown_ = ( Costmap2D(ma).getDefaultValue() == 0 ) ? false : true;

The ma object that's passed into the functions that have this line is
defined as a constant.  The getDefaultValue function itself is not
declared as a constant function.  Unfortunately, this means that we
can't invoke the 'ma' objects call. So, we make a temporary
non-constant copy, and invoke the call with the new object.