costmap_3d: switch FCLFloat to double
From testing, using doubles is slightly faster for the distance
checks, although a bit slower for collision. Most clients will be
querying distance the most heavily (to optimize distances in path
processing), and the accuracy is needed when operating in a fixed
frame, as the querying may be relatively far from the origin.