costmap_3d: add non-const version of updateCells
Add a version of updateCells that takes mutable pointers to the
value and bounds trees and alters their depth to match the layer's
costmap depth.

Without adding a new interface, the old interface can only work
with trees of the same depth, as only the const version of
setTreeValues can be called (which will error-out when trees of
differing depth are passed).