costmap_3d: use unsigned ints in 2d conversion
The 3D to 2D conversion plugin was using signed integers when
calculating the 2D map index from the 3D index. This does not work
correctly for 32-bit 3D indexes. Instead, use unsigned integers
everywhere for the conversion, and make use of octomap's coordToKey
for finding the 2D map's origin in the 3D index space.