costmap_3d: fix calculation of signed distance
When the signed distance computation was negative and the center of
the octomap box was inside the mesh, the signed distance was not
negated before subtracting the diameter of the box. This results in
penetration depths that are too deep when on the boundary. This can
cause major problems in solvers that use the derivative by creating a
minima that would capture the mesh inside an obstacle. Fix this by
always negating the signed distance for an interior collison and then
subtracting the diameter.