Navfn: Make params dynamic reconfigurable
* allow_unknown
* default_tolerance

I did not bother with plannery_window_x (& _y) because these parameters
appear to be unused by the code.

Also, I left visualize_potential non-dynamic, preferring instead to make
it only publish if there is an active subscriber, which would allow this
to be enabled all the time without penalty.