costmap_3d: add reuse results, last cache thread local
Add a flag to the query functions to reuse the results from the last
call directly. Make the last cache a thread-local cache. This way,
a caller can call a query once at a pose, then perturb it slightly,
and set the reuse flag to get a very fast answer to the small
perturbation. By making the last cache thread-local, the caller is
guaranteed to reuse the last call it made.