costmap_3d: add interface to update costmap in buffered query
Add an interface to update which costmap a buffered query is
attached to. This allows the robot-model dependent data to be
saved (such as the interior collision LUT) while swapping the
map contents. Add an interface to Costmap3DROS for planners
to use to update a buffered query with a copy of the latest
layered costmap.