costmap_3d: add relative error to distance queries
Add abililty to specify relative error to distance queries.
For instace, w/ relative error .05, the returned distance will be
within 95% of the correct distance (for a positive distance).

Using a relative error greatly reduces the amount of time needed
to query by allowing large sections of the search space to be
pruned. Imagine an upgright cylinder bounding volume for the mesh
and many octomap cells in a vertical stack like a wall. By
allowing a small relative error, many of the boxes will not be
close enough to the main bounding volume to improve the answer
by more than the relative error, allowing us to stop early.