costmap_3d: save and update buffered queries for srvs
Instead of creating a new buffered query object for each buffered
query via the service or action server interfaces, save the
created queries in a map indexed by mesh+padding. If we already
have a buffered query matching an incoming query, reuse it by
simply updating the costmap associated with it. This saves any
other ancilliary setup steps (like calculating LUTs from the mesh).

As a part of this, the plain ROS service and the action server
now serialize requests from each source to prevent from updating
the buffered query on each other (which might change the costmap
they are querying against mid-query).

This isn't an issue as any serious query user should be using
the action server interface, as it is on its own thread. The
ROS service will block all the sensor callbacks while executing
its query preventing the costmap from updating, which is a problem.