costmap_2d: add method to get the transform listener
For a global planner to create its own private Costmap2DROS,
it needs a tf::TransformListener. However, it is only passed
a pointer to the global costmap's Costmap2DROS. Therefore,
add a method to get the transform listener so it can be
passed into the constructor of a Costmap2DROS. This allows
a global planner to create its own private costmap.