navfn: do not clear the robot cell
It is incorrect to clear the "robot cell" as the planner may be
called with some future position for the starting pose of the plan.
Also, clearing just one cell does not handle clearing the entire
footprint. Footprint clearing is outside the scope of the planner
and belongs in the costmap itself.