costmap_3d: import fcl via cmake
The previous method used pkgconfig but then only used the paths to pass
into find_library and find_path to recreated the include path and
library path. Using pkgconfig directly instead of like this would have
been better but since changes needed to be made using the cmake file is
even better. This allows us to just add a dependency on the fcl target
and all of the needed compile and link options will be used.

The previous method was actually missing out on a number of flags that
are passed from fcl. We happened to be getting lucky for a while but it
seems some changes to apt packages may have broken this. Specifically by
using find_library it only found the path to the fcl library and linked
against it but it didn't add the additional required libraries to link
against that fcl also specified. In this specific case it was not
linking against libccd even though this is needed in order to use fcl.
It was also missing out on additional include directories.

One caveat of using the cmake package is that catkin_package doesn't
have a great way to add imported dependencies like this. I would
actually argue that the catkin_package call is actually lacking since it
doesn't take into account the link directories needed to find the
libraries in question but there is nothing to be done about that. In
order to extract the needed flags for catkin_package we use the
get_target_property() call to populate the used variables.