costmap_3d: add link library paths
If packages need to add specific link paths we need to honor them.
Specifically this is needed by octomap. Currently it was getting by
because the find_library call for fcl was adding the link paths that it
needed but we really need to add these so we can change the way fcl is
added and so that we aren't relying on indirect methods for making
things work.