costmap_3d: add exact signed distance query mode
Add a new exact signed distance mode to queries. This mode will find the
exact signed distance by not stopping the distance query when a
collision is found, but exhausting all interior collisions to find the
deepest. The normal (relaxed) signed distance mode stops as soon as a
collision is found, but still allows the caller to keep derivates
consistent by use of the reuse last result option. Some users of costmap
queries (such as recovery planners) want to know the *exact* penetration
depth for any particular pose of the mesh, and can do so now with this
new exact signed distance query mode.

Note: to truly get exact penetration depth also requires setting the
relative error to zero for the query.

If only the penetration depth is needed (in other words, if
non-collision distance is irrelevant to the caller) the query can be
sped up by setting the min_distance via the distance options to be just
above zero. This speed up is amortized to nothing by the caches, so if
the query will result in a lot of cache hits over time, the speed up is
negligible. But if only a few queries are to be issued, the speed up
could be considerable.

Finally, exact signed distance does not imply exact distance for
non-collisions which are still short-circuited by the milli and micro
caches. Those can be disabled by creating a query object with the milli
and micro cache distance thresholds set to the maximum value (preventing
them from being directly used to calculate distance on a hit). Such
precise distances in non-collision cases should not be necessary for
normal use cases.