costmap_3d: binPose: fix truncation, cayley scale
The binPose function was using static cast to a floating point type
to truncate the number. This always goes towards zero, which makes
the bin at the origin twice the size it should be. Because it is
unlikely to be operating around the origin, this bug went unnoticed.
Instead, use std::floor which is still very fast on modern
architectures and keeps all the bins the same size.

Furthermore, binPose was binning the cayley transform as if its units
are radians. However, the spacing of the transform works out to
almost excatly 1/4 of a radian near zero rotation, so scale the
binning factor by 4 to get roughly the correct spacing in radians.
The previous code would result in incorrectly spaced rotation bins,
causing them to be roughly 4 times bigger than expected. The
consequence of this is worse performance, as although the cache is
hit more-often, the bound is poorer, or, in the case of the
milli-cache, more inaccurate results as the cache may be used for a
significantly larger rotation than intended.