costmap_3d: query: make reuse results lockless
Handle reuse results first before acquiring any locks.
It is already up to the caller using reuse results to ensure the
costmap hasn't updated in-between calls, so it is pointless to
acquire the lock to check. Because reuse results only relies on
shared data that can only change curing a costmap update and on
thread-local storage this is completely safe.