costmap_3d: query: fix read locking
The original implmentation mis-interpreted boost upgrade locks as being
shared. They are not. Only one thread may own an upgrade lock at a time.
Instead, one must use shared locks on the upgrade mutex for shared
(read-only) access. Fix this by reorganizing the code a bit to only
acquire a shared lock towards the beginning of the costmap query to
check if shared state must be updated (such as the costmap pointer or
nonlethal costmap pointer). Then, if either of those might need to be
updated, the shared lock is dropped, an upgrade lock is acquired, the
state is re-checked and updated if necessary (after upgrading to an
exclusive lock) then the upgrade lock downgraded to a shared lock.

This keeps the fast-paths shared via the shared mutex and the only time
shared locks will need to block is during cache updates at the end. This
greatly improves the paralleliztion of costmap queries (as before only
the octree solver could run in parallel, and all the fast paths would
block each other on the upgrade lock).

Note: std::shared_mutex is avoided as its implementation is allowed to
starve writers by waiting forever for no shared locks to be held. The
boost implementation guarantees that locking an upgrade_mutex blocks any
new readers (shared locks) from succeeding until after the exclusive
lock has been acquired and released. This prevents writers from being
starved out by readers.