costmap_3d: query: grab exclusive lock in updateMeshResource
The updateMeshResource modifies a lot of internal state and should be
holding an exclusive lock while doing so. As of right now, no one is
modifying the mesh after creating a query, but for correctness it should
be locking in case a caller is ever added after the query has been
created.