costmap-3d: query: make TLS cache by copy
Make the thread-local store cache store cache entries by copy and not
be reference. It is not safe to hold them by reference as on fast
micro and milli hits the TLS cache needs to be updated, and referring
to a milli or micro cache entry is not thread-safe, as they may be
deleted. Since the TLS cache is used without any locks it is
necessary to copy the cache entry into the TLS. This is not expensive
as distanc3 cache entries are fairly small. This change has the plus
side of not needing the write lock to update the TLS cache anymore.