costmap-3d: query: skip cache update if cached neg. dist.
If there was a cached collision (negative distance), skip updating
the shared caches (do update the TLS cache). It is expensive to get
the write lock. Because there was already a cached collision, any
future queries that would hit this entry should also hit the
collision as well, so the cost of updating outweighs any benefit.