costmap-3d: query: re-organize read locking
Re-organize the read locking to use RAII nearly everywhere and to only
hold the read locks when absolutely necessary. Cache entries are copied
to the stack as the GJK solver in FCL can take a microsecond or so to
solve the triangle/box problem. Keeping the read locking to a minimum
makes cache updates faster.