costmap-3d: query: fix query region caches
The distance caches were being shared between the different query types
and regions. This is wrong as the non-lethal tree is different from the
lethal tree, and the caches will be wrong for the various
regions-of-interest. Also, the caches were being updated with the most
recent hit, however this invalidates any TLS entries pointing to micro
or milli cache entries from the fast-path hits. To fix this, do not
update or alter a cache entry, only add an entry if a previous entry did
not exist. Lastly the TLS was not being checked in some cases by a
previous commit, which breaks TLS as it must be checked each time to
ensure the TLS of the current thread is still valid.