costmap_3d: query: make reuse results lockless
Handle reuse results first before acquiring any locks. This speeds up
parallelism and increases performance of parallel queries. It is
already up to the caller using reuse results to ensure the costmap
hasn't updated in-between calls. However, other threads may modify
the cache entries themselves so to be completely lockless the
distance cache entry needs to be copied into the TLS. This is not
expensive as distance cache entries are fairly small. It is necessary
to be able to invalidate an entry, which is easily done by
invalidating the shared pointers to the octomap box and mesh triangle
in the entry.