costmap_3d: query: improve fast-path performance
Improve the fast-path performance for micro and milli cache hits by
testing the threshold against both the *calculated* distance and the
stored distance. For this to work, the milli cache needs to be
invalidated differently during costmap updates. Instead of storing
the milli cache distance threshold, just set the stored distance to
infinity and check if the distance is finite during the milli cache
check. This helps as the threshold is valid either way and planners
that are querying will tend to adjust poses *away* from obstacles.
Also add collisions to the fast path when exact distance mode is not
enabled (as all that needs to be returned is the first collision).