costmap_3d: query: switch to std::shared_timed_mutex
Ditch boost::upgrade_mutex in favor of std::shared_timed_mutex. While
the Boost upgrade mutex tries to be fair, and has upgrade locking,
its implementation is slower, and even causes slowness on the read
side as shared locks are merely emulated using a normal mutex. Switch
over to using C++14's std::shared_timed_mutex. If all code using
Costmap3D is converted to C++17 (such as in a future ROS release)
this mutex could be switched to just std::shared_mutex as the timed
feature is not used.