costmap_3d: octree solver: move to sphere-OBB signed distance
Instead of using RSS-RSS distance for guiding and pruning the octree
solver, use sphere-OBB signed distance. Model each octomap volume as
the bounding sphere of that volume. Calculating sphere-OBB signed
distance can be done very quickly and also accurately (over 8 times
faster than RSS-RSS). Also, the process FCL uses to fit the bounding
volumes first fits an OBB and then simply fits an RSS to that OBB,
which means the OBB will always be of smaller volume, and thus prune
the search better. Because many thousands of octomap-mesh bounding
volume distances are calculated for every query this is a significant
performance improvement.