costmap_3d: octree solver: argsort distances
Instead of using a custom nested loop to find the next minimum
distance, use an internal optimized index sorter. This is significantly
faster than the old solution as it has fewer loops and fewer branches
and uses fewer comparisons overall. For comparison, the optimized
sorter can sort random sequences of 2-8 distances in 10ns on average,
where std::sort on the same sequences takes 34ns on average. While this
may seem like a negligible improvement, the sorting must take place in
every octree recursion, so is performed tens of thousands of times per
planning cycle. This optimization therefore results in a measurable (but
modest) performance increase of about 2-4% on the planning cycle. This
speedup is achieved by using optimal sorting networks and by packing
the distance and index into a single 64-bit integer and using
branchless swap-if-greater than constructs using ternary operators.
This emits cmov instructions on x86 (and would emit conditional moves
on other architectures as well) which are branchless.