costmap_3d: fix test_octree_solver
The old test did not work at all, even though it passed. This was
because the free threshold had not been set on the FCL octree, so the
entire octree was thought to be free space, so all distance queries
just did nothing. Also, the FCL broadphase check had its TF's
backwards and was not applying the region of interest to the
generated boxes. After fixing those things, it was discovered that
somewhere in FCL the broadphase check is broken when the boxes are
rotated, so instead of transforming both the octree and mesh, leave
the octree at identity transform. Finally, the old 'aisles.bt' map
had the aisles shifted off more than 30 meters from the origin, so
the checks had no collisions or close objects which is unrealistic.
So, generate a new aisles.bt map with the aisles in the center of the
map. Change the x/y extents from +/- 2m to +/- 10m to vary over the
maps bounds better.