costmap_3d: octree solver: drop const from distance
Make the distance method non-const and drop 'mutable' from members. It
made little sense to mark the distance function and functions it calls
'const' when the object is being mutated anyway, and communicates the
wrong thing to callers. Drop the const to indicate that the distance
method changes class instance state (so is inherently non-thread safe).