costmap_3d: octree solver: move roi to obj
Instead of passing roi around, move it into the object.

Make whether or not to check the ROI part of the template for the
internal octree recursion. With this change, when check ROI is false
the code does not have to check the pointer at all saving a test and
branch. This is helpful as when the branch switches from being taken
(when there is an ROI high in the tree) to not-taken it will be
mis-predicted several times resulting in a perfromance hit.