costmap_3d: octree solver: make child BV compute branchless
Make a version of FCL's computeChildBV called computeChildMinMax that
is branchless and inline and avoids constructing the AABB until
necessary. The octree descent computes child BVs on a fast-path, so
it needs to run as fast as possible. Avoiding constructing the AABB
until necessary saves time initializing 8 AABBs on every pass and
instead only initializes the ones that will actually get used, and
then only to the values they need.