costmap_3d: octree solver: fast-forward single children
If the octree has only a single child, fast forward instead of
recursing. Also, calculate the children number and pointers during
this check and re-use them below. This also keeps the stack growth
optimal as there are fewer recursive calls, and only space for the
distances for the children present gets allocated on the stack. To
make this work, make the bounding volume argument to the octree
recursive function mutable. While this may seem like a small
optimization, in practice it saves 3-5% of the work.