costmap_3d: query: option for direct cache use below thresh
Add options to directly use milli or micro cache distances when above or
below the distance threshold. Previously the results above the threshold
were always used, that is now optional and enabled by default. Add a new
option to reuse the milli or micro cache distances when below the
threshold by first checking if there is a collision (by limiting
min_distance to numeric_limits::min()) and if there is not, directly
using the result. This makes slow-milli cache hits much faster (nearly
twice as fast on average during testing) at the expense of introducing
potentially large relative errors (by missing a closer object within the
resolution of the cache), but it still collision-safe.