costmap_3d: octomap_server_layer_3d: add queue size param
Add a queue size parameter for the octomap_server_layer_3d for the map
update topic named map_update_queue_size. The default is 30 to allow for
up to 1 second of delay processing the map of a 30Hz sensor map. It is
important to not miss an update as having to resubscribe to get a full
update can take significant time.