costmap_3d: query: del reuse_results, thread_local
Remove the reuse_results API, and remove all thread_local variables.
Replace tls_last_cache_entries_ with a last_distance_cache_entries_
protected by the distance cache mutex. Provide the same functionality
as reuse_results by adding an optional DistanceResult return structure
which can be reused in a future call to calculateDistance. The
reuse_results was rather hackish as it depended on maintaining the
previous call's results. The thread_local last cache was brittle and
hard to maintain and provided no added benefit over just using member
variables to track the last cache entry for each of the types of
queries. The results of these changes are to make the API cleaner, and
to make the Costmap3DQuery object easier to maintain, with no loss in
performance. In fact, there may be a slight gain on average due to not
needing to check thread_local variables (which often generate function
calls to see if they need to be initialized).