costmap_3d: fix unsubscribe deadlocks
Calling `shutdown` on a subscription synchronizes on all outstanding
callbacks from the subscriber. If those callbacks grab a lock, it is
incorrect to call unsubscribe while holding said lock. It is safe to
unsubscribe without holding the lock as the subscribers are
thread-safe. After the unsubscribe completes, it is guaranteed that
no callbacks are running (or will be scheduled) from the subscriber.

Move all places where unsubscribe was called outside of holding the
locks that are acquired inside their respective callbacks. Otherwise,
deadlocks ensue if using a multi-threaded spinner.